home, list, details pages added with functinoality

+++ pgai/empstat/models.py
@@ -6,7 +6,7 @@ from django.db import models
 class Employee(models.Model):
     name = models.CharField(max_length=100)
     department = models.CharField(max_length=50)
-    salary = models.DecimalField(max_digits=10, decimal_places=2)
+    salary = models.IntegerField()
 
     def __str__(self):
         return self.name
